Add files via upload
This is a Neural Network example, which approximates a x^2 function (Check the Universal Approximation Theorem). The code does not use any exisiting ML library and hence is hence a good way to learn the basics of Neural Network. I have used Sigmoid activation function and a gradient descent method to train the neural network.